refactor(chat): split sidebar into composables
Extract the 2.5k-line WindowSideBar script into seven sidebar
composables (workspace groups, pin flight, list auto-fill,
session shortcuts, group reorder, workspace actions, remote
control) and replace raw window/document API usage with VueUse
primitives (useEventListener, useDocumentVisibility,
useResizeObserver, useTimeoutFn, usePreferredReducedMotion).

Also lands in the sidebar itself:
- duplicate workspace labels rendered with minimal parent
  context over the visible group set, full normalized path on
  the group header title and an sr-only span (#2121)
- a subtle chat-section card background plus a divider above
  the workspace section so the two lists read separately

Template and DOM structure are unchanged. The two tests that
poked component internals via wrapper.vm now go through real
window events and the poll timer instead.
